// src/auth/enums/auth-error-codes.enum.ts
export enum AuthErrorCode {
  AUTH_OTP_INVALID = 'AUTH_OTP_INVALID',
  AUTH_OTP_EXPIRED = 'AUTH_OTP_EXPIRED',
  AUTH_OTP_TOO_MANY_ATTEMPTS = 'AUTH_OTP_TOO_MANY_ATTEMPTS',
  AUTH_OTP_RATE_LIMITED = 'AUTH_OTP_RATE_LIMITED',
  AUTH_SESSION_EXPIRED = 'AUTH_SESSION_EXPIRED',
  AUTH_REFRESH_TOKEN_INVALID = 'AUTH_REFRESH_TOKEN_INVALID',
  AUTH_CREDENTIALS_INVALID = 'AUTH_CREDENTIALS_INVALID',
  AUTH_USER_BLOCKED = 'AUTH_USER_BLOCKED',
  AUTH_USER_ARCHIVED = 'AUTH_USER_ARCHIVED',
  AUTH_USER_EXISTS = 'AUTH_USER_EXISTS',
  AUTH_USER_NOT_FOUND = 'AUTH_USER_NOT_FOUND',
  INTERNAL_SERVER_ERROR = 'INTERNAL_SERVER_ERROR',
}